+++ apps/v4/examples/base/input-group-block-start.tsx
@@ -1,9 +1,11 @@
+import { CopyIcon, FileCodeIcon } from "lucide-react"
+
 import {
   Field,
   FieldDescription,
   FieldGroup,
   FieldLabel,
-} from "@/examples/base/ui/field"
+} from "@/styles/base-nova/ui/field"
 import {
   InputGroup,
   InputGroupAddon,
@@ -11,8 +13,7 @@ import {
   InputGroupInput,
   InputGroupText,
   InputGroupTextarea,
-} from "@/examples/base/ui/input-group"
-import { CopyIcon, FileCodeIcon } from "lucide-react"
+} from "@/styles/base-nova/ui/input-group"
 
 export function InputGroupBlockStart() {
   return (